% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-groups.R
\name{get_resource_group}
\alias{get_resource_group}
\title{Get Resource Group by Name}
\usage{
get_resource_group(x, name)
}
\arguments{
\item{x}{Azure subscription.}

\item{name}{Name of resource group.}
}
\description{
Get Resource Group by Name
}
