% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-groups.R
\name{list_resource_groups.default}
\alias{list_resource_groups.default}
\title{List of All Resource Groups in All Subscriptions}
\usage{
\method{list_resource_groups}{default}(x, ...)
}
\arguments{
\item{x}{Azure subscription.}

\item{...}{Addition parameters for
\code{\link[AzureRMR]{az_subscription}$list_resource_groups} S6 method.}
}
\value{
Named list of resource groups, R6 environments. The name of the group
concatenates the subscription identifier with the resource group name with
full-stop delimiter.
}
\description{
Asks for all subscriptions, since none supplied. Lists resource groups for
each one then concatenates the lists. The result becomes a combined list of
resource groups. The subscription identifier remains accessible from each.
}
\examples{
\dontrun{
purrr::keep(list_resource_groups(), ~ .x$name == "rg")
}
}
