% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-groups.R
\name{list_resource_groups}
\alias{list_resource_groups}
\title{List Resource Groups}
\usage{
list_resource_groups(x, ...)
}
\arguments{
\item{x}{Azure subscription, or \code{NULL} for all available subscriptions.}

\item{...}{Addition parameters for
\code{\link[AzureRMR]{az_subscription}$list_resource_groups} S6 method.
Azure supports \code{filter} by tag name and value; also \code{top} for limiting the
number of resource groups.}
}
\value{
List of resource groups.
}
\description{
List Resource Groups
}
