% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{list_subscriptions}
\alias{list_subscriptions}
\title{List Azure Subscriptions}
\usage{
list_subscriptions(x, ...)
}
\arguments{
\item{x}{Azure resource manager, an \code{az_rm} instance.}

\item{...}{Additional parameters for
\code{\link[AzureRMR]{az_rm}$list_subscriptions}.}
}
\value{
List of subscriptions (specifically \code{az_subscription} R6
instances) named for their subscription identifier.
}
\description{
List Azure Subscriptions
}
