% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-groups.R
\name{list_resource_groups}
\alias{list_resource_groups}
\title{List Resource Groups}
\usage{
list_resource_groups(x, ...)
}
\arguments{
\item{x}{Azure subscription.}

\item{...}{Addition parameters for
\code{\link[AzureRMR]{az_subscription}$list_resource_groups} S6 method.}
}
\value{
List of resource groups.
}
\description{
List Resource Groups
}
